Extracted director data
